% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/maxSiteTime.R
\name{maxSiteTime}
\alias{maxSiteTime}
\title{Calculate count maxima per site and time}
\usage{
maxSiteTime(x, time)
}
\arguments{
\item{x}{An object of class OrnithoData}

\item{time}{A time space in which to return maxima, either "decade" or "pentade", the default is "decade"}
}
\value{
Returns a data.frame with count maxima per site
}
\description{
Extracts the count maxima per decade or pentade of a specific site
}

