% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/obsSiteTime.R
\name{obsSiteTime}
\alias{obsSiteTime}
\title{Calculate number of observations per site and time}
\usage{
obsSiteTime(x, time)
}
\arguments{
\item{x}{An object of class OrnithoData}

\item{time}{A time space in which to return maxima, either "decade" or "pentade", the default is "decade"}
}
\value{
Returns a data.frame with count maxima per site
}
\description{
Counts the observations per site and decade/ pentade
}

