\name{ornitho2natis}
\alias{ornitho2natis}

\title{
Conversion from ornitho data to natis data
}
\description{
A list of \code{ornithoSpec} objects is converted into a data frame and vice versa.
}
\usage{
ornitho2natis(x, file)
}
\arguments{
  \item{x}{
a list, whose elements are objects of class \code{\link{ornithoSpec}}.
}
\item{file}{a charater string representing the directoriy to write a natis dfb file}.
}

\references{
The German \bold{ornitho} website: \url{http://www.ornitho.de/}
}
\author{
Simon Thorn
}


