\name{projOrnithoMap}
\alias{projOrnithoMap}
\title{
Map Projection and Datum Transformation
}
\description{
Map projection and datum transformation for objects of class \code{\link[=ornithoMap-class]{ornithoMap}}.
}
\usage{
projOrnithoMap(map, projection)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{map}{
an object of class \code{\link[=ornithoMap-class]{ornithoMap}}.
}
  \item{projection}{
an object of class \code{\link{CRS}}.
}
}
%\details{}
\value{
an object of class \code{\link[=ornithoMap-class]{ornithoMap}}.
}
%\references{}
\author{Christoph Heibl}
\seealso{
\code{\link{ornithoMap}}                                  
}
%\examples{}