\name{rornitho-internal}

\alias{attribute}
\alias{crop.grid}
\alias{crop2polygon}
\alias{def}
\alias{extract.coords}
\alias{grid.in.polygon}
\alias{missingObservations}
\alias{obsMaxSiteTime}
\alias{plotOrnithoMTB}
\alias{quarter2full}
\alias{setAttributes}
\alias{speclist}
\alias{subsetOrnitho}

\title{Internal RORNITHO Functions}
\description{Internal \pkg{rornitho} functions.}
\note{
These are internal functions in \pkg{rornitho} and are not intended to be called by the user.
}
\keyword{internal}
